move specific log functions to other modules
