+++ src/managers/log.h
@@ -1,19 +1,16 @@
 #ifndef ACE_MANAGER_LOG_H
 #define ACE_MANAGER_LOG_H
 
+#include "config.h"
+
 #include <stdio.h> // fopen etc
 #include <string.h> // strlen etc
 #include <stdarg.h> // va_list etc
 #include <clib/exec_protos.h> // Amiga typedefs
 #include <clib/graphics_protos.h> // Amiga typedefs
 
-#include "managers/timer.h"
-
-
-
 
 #ifdef GAME_DEBUG
-
 
 
 /* Functions - general */
@@ -31,36 +28,11 @@ void logWrite(
 
 /* Functions - block logging */
 
-void logBlockBegin(
-	IN char *szBlockName,
-	IN ...
-);
-void logBlockEnd(
-	IN char *szBlockName
-);
+void logBlockBegin(char *szBlockName,...);
+void logBlockEnd(char *szBlockName);
 
 
 /* Functions - struct dump */
-
-void _logUCopList(
-	IN struct UCopList *pUCopList
-);
-void _logBitMap(
-	IN struct BitMap *pBitMap
-);
-
-/*# define logOpen() _logOpen()
-# define logClose() _logClose()
-# define logPushIndent() _logPushIndent()
-# define logPopIndent() _logPopIndent()
-# define logWrite(...) _logWrite(__VA_ARGS__)
-
-# define logBlockBegin(...) _logBlockBegin(__VA_ARGS__)
-# define logBlockEnd(szBlockName) _logBlockEnd(szBlockName)
-*/
-
-# define logUCopList(pUCopList) _logUCopList(pUCopList)
-# define logBitMap(pBitMap) _logBitMap(pBitMap)
 
 #else
 
@@ -70,12 +42,9 @@ void _logBitMap(
 # define logPopIndent()
 # define logWrite(char *szFormat,...)
 
-# define logBlockBegin(char *szFormat...)
+# define logBlockBegin(char *szFormat,...)
 # define logBlockEnd(szBlockName)
 
-
-# define logUCopList(pUCopList)
-# define logBitMap(pBitMap)
 #endif
 
 #endif /* end of ACE_MANAGER_LOG_H */
