+++ src/managers/performance.c
@@ -1,9 +1,9 @@
 #include "performance.h"
+#ifdef GAME_DEBUG
 
 #include "managers/timer.h"
 #include "managers/memory.h"
 #include "managers/log.h"
-
 
 
 /**
@@ -40,15 +40,15 @@ void logAvgBegin(tAvg *pAvg) {
  *
  */
 void logAvgEnd(tAvg *pAvg) {
-	// Calculate timestamp
+	/* Calculate timestamp */
 	pAvg->pDeltas[pAvg->uwCurrDelta] = timerGetDelta(pAvg->ulStartTime, timerGetPrec());
-	// Update min/max
+	/* Update min/max */
 	if(pAvg->pDeltas[pAvg->uwCurrDelta] > pAvg->ulMax)
 		pAvg->ulMax = pAvg->pDeltas[pAvg->uwCurrDelta];
 	if(pAvg->pDeltas[pAvg->uwCurrDelta] < pAvg->ulMin)
 		pAvg->ulMin = pAvg->pDeltas[pAvg->uwCurrDelta];
 	++pAvg->uwCurrDelta;
-	// Roll 
+	/* Roll */
 	if(pAvg->uwCurrDelta == pAvg->uwCount)
 		pAvg->uwCurrDelta = 0;
 }
@@ -63,14 +63,15 @@ void logAvgWrite(tAvg *pAvg) {
 	char szMin[15];
 	char szMax[15];
 	
-	// Calculate average time
+	/* Calculate average time */
 	for(i = pAvg->uwCount; i--;)
 		ulAvg += pAvg->pDeltas[i];
 	ulAvg /= pAvg->uwCount;
 	
-	// Display info
+	/* Display info */
 	timerFormatPrec(szAvg, ulAvg);
 	timerFormatPrec(szMin, pAvg->ulMin);
 	timerFormatPrec(szMax, pAvg->ulMax);
 	logWrite("Avg %s: %s, min: %s, max: %s\n", pAvg->szName, szAvg, szMin, szMax);
 }
+#endif /* end of GAME_DEBUG */
